create "Services" folder for a service-like class. Rename JsonDbCotroller to JsonRepository. Move it to "Services" folder
